refactor(tui): extract card detail sections into reusable components
- Create card_detail_sections module with builders for title, metadata, description, and sprint logs
- Replace 100+ lines of duplicated rendering code with helper functions
- Reduce render_card_detail_view from 310 to 150 lines through functional composition
- Improve testability and maintainability of section rendering logic
